Fix unit test error
Fix the intensional unit test error

+++ Test1/Algebra.cs
@@ -4,8 +4,8 @@
     {
         public long Add(int num1, int num2)
         {
-            //return (long)num1 + (long)num2;
-            return num1 + num2;
+            return (long)num1 + (long)num2;
+            //return num1 + num2;
         }
 
         public long Sub(int num1, int num2)
